[Instrumented libraries] Fix libxkbcommon0 build
In order for MSAN to work, all DSOs that get loaded for a process
(the main executable and its shared libraries) must all be instrumented
with MSAN if any pointers are to be passed across DSO boundaries.  This
allows MSAN to know, for example, that a buffer that was allocated and
initialized by a shared library and passed back to Chrome, was in fact
initialized.

This means that in MSAN configs, package tests are expected to fail
since they're not run with dependent instrumented libraries -- instead
they're run with the system libraries since otherwise the build would
fail as the compiler itself wasn't built with MSAN.  The fix is to
simply add nocheck/notest to skip the tests.  I've also added nodoc
to skip generating documentation and nostrip to leave debug symbols
in case they're useful.

R=jperaza

Bug: None
Change-Id: I1274e4355f5a49440497351049294f5064527fb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2669842
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Joshua Peraza <jperaza@chromium.org>
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Joshua Peraza <jperaza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#849908}
GitOrigin-RevId: a40143305163b428375c1f6a7da6bb2a9ea09a64